feat(frontend): implement DSL converters and mock API layer (#63)
Introduces the initial implementation of workflow creation and
management. It includes a refactored workflow canvas, a table view to
list existing workflows, and the underlying frontend architecture to
support these features.

Also added a Mock API, `api-mock.ts` wired to the frontend that provides
hardcoded data to help with testing.

The workflows list page is currently a static bare-bones component, it
would ideally require the following to be done to be ready:
* **Workflows API:**
       - `GET /workflows`: To fetch the list of all workflows.
- `GET /workflows/{id}/graph`: To fetch the graph for a specific
workflow.
       - `POST /workflows`: To create a new workflow.
       - `PUT /workflows/{id}/graph`: To save a workflow's graph.
- `PUT /workflows/{id}/status`: To activate or deactivate a workflow.
       - `DELETE /workflows/{id}`: To delete a workflow.
* **Database Schema:**
- A workflows table is needed to store workflow metadata and the
workflow graph (as JSON).

The canvas refactor disables the save functionality, we will re-enable
this when the workflows list page is ready to wire it to the users' list
of workflows. It also adds JSON parsing and export, you can find sample
workflow JSONs under `samples/`, you can also manually create your own,
export and paste (CTRL+V) onto the canvas for testing.